8index.html
Here’s a summary of the updates:

Enlarged Header: The header section is now slightly larger and more prominent.
Optional Image Title: I've added a section in the code for an image logo. It's currently commented out. To use it, you'll just need to add your image file and swap which line is commented out, as explained in the code comments. This image logo is linked to priyamalok.me.
Larger Navigation Menu: The font size for the "Home," "Projects," and "Musings" links has been increased.
Updated Links: All navigation and content links have been updated to the new URLs you provided.